#160809125: Pequenas modificações na função de exclusão de contas

+++ src/app/admin/deposito/edit-deposito/edit-deposito.component.ts
@@ -1,13 +1,14 @@
 import { Component, OnInit } from '@angular/core';
 import { Observable, Subject } from 'rxjs';
 
+import { EventoService } from '../../../services/evento.service';
 import { NotificacaoService } from '../../../services/notificacao.service';
 import { DepositoService } from './../../../services/deposito.service';
 
 @Component({
   selector: 'ngx-edit-deposito',
   templateUrl: './edit-deposito.component.html',
-  styleUrls: ['./edit-deposito.component.scss']
+  styleUrls: ['./edit-deposito.component.scss'],
 })
 export class EditDepositoComponent implements OnInit {
   form: any = {};
@@ -16,7 +17,11 @@ export class EditDepositoComponent implements OnInit {
   contasID: Observable<any>;
   eventoResolver: any = [];
 
-  constructor(private depositoservice: DepositoService, private notific: NotificacaoService) { }
+  constructor(
+    private depositoservice: DepositoService,
+    private notific: NotificacaoService,
+    private eventoservice: EventoService,
+  ) { }
 
   ngOnInit() {
     this.contas = this.depositoservice.getContaDeposito();
@@ -36,8 +41,18 @@ export class EditDepositoComponent implements OnInit {
   }
 
   excluirConta() {
-    this.depositoservice.removeContaDeposito(this.eventoResolver.id);
-    this.notific.ngxtoaster('Conta excluida com sucesso!', '', true);
+    this.eventoservice.getcontasDepositoId(this.eventoResolver.id).subscribe(res => {
+      if (res.length > 0) {
+        this.notific.ngxtoaster(
+          'Ainda temos eventos com essa conta registrada!',
+          'Altere estes eventos para poder deletar a conta',
+          false,
+        );
+      } else {
+        this.depositoservice.removeContaDeposito(this.eventoResolver.id);
+        this.notific.ngxtoaster('Conta excluida com sucesso!', '', true);
+      }
+    });
   }
 
 }
